Convert to Qt4 - rename some QWMatrix to QMatrix Fix multipleduplicate gui bug with gap/shift signal and label change Other minor changes
git-svn-id: svn://scribus.net/trunk/Scribus@10030 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scribus.cpp
@@ -3448,7 +3448,7 @@ bool ScribusMainWindow::slotPageImport()
 		else
 		{
 			startPage = doc->currentPage()->pageNr() + 1;
-			if (static_cast<uint>(nrToImport) > (doc->DocPages.count() - doc->currentPage()->pageNr()))
+			if (nrToImport > (doc->DocPages.count() - doc->currentPage()->pageNr()))
 			{
 				qApp->changeOverrideCursor(QCursor(Qt::ArrowCursor));
 				int scmReturn=ScMessageBox::information(this, tr("Import Page(s)"), "<qt>" +
